requirement.txt
streamlit
openai
llama-index
pathlib
nltk
llama-index-readers-file
python-dotenv
